[QOL-7906] refactor controller
- split Flask vs Pylons in their own directory, like the plugin split
